requirements.txt
numpy==1.26.2
pandas==2.1.3
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0